+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/engine_functions/logs_components/loki/lokig_config.go
@@ -54,8 +54,8 @@ type Configs struct {
 	From        string `yaml:"from"`
 	Store       string `yaml:"store"`
 	ObjectStore string `yaml:"object_store"`
-	Schema      string `yaml:"schema"`
-	Index       Index  `yaml:"index"`
+	Schema string `yaml:"schema"`
+	Index  Index  `yaml:"index"`
 }
 
 type SchemaConfig struct {
@@ -137,7 +137,7 @@ func newDefaultLokiConfigForKurtosisCentralizedLogs() *LokiConfig {
 		},
 		LimitsConfig: LimitsConfig{
 			RetentionPeriod: limitsRetentionPeriod,
-			AllowDeletes: allowDeletes,
+			AllowDeletes:    allowDeletes,
 		},
 		Analytics: Analytics{
 			ReportingEnabled: analyticsEnabled,