+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/engine_functions/logs_database/loki/loki_container_config_provider_factory.go
@@ -1,7 +1,0 @@
-package loki
-
-func CreateLokiContainerConfigProviderForKurtosis() *LokiContainerConfigProvider {
-	lokiConfig := newDefaultLokiConfigForKurtosisCentralizedLogs()
-	lokiContainerConfigProvider := NewLokiContainerConfigProvider(lokiConfig)
-	return lokiContainerConfigProvider
-}